requirements.txt
Click==7.0
Flask==1.1.1
Flask-Caching==1.7.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.17
peewee==3.10.0
pkg-resources==0.0.0
protobuf==3.9.1
python-dotenv==0.10.3
six==1.12.0
uWSGI==2.0.18
Werkzeug==0.15.5